// src/app/core/models/company.offer.new.model.ts
export class CompanyNewOfferModel {
    image_one: string;
    image_two: string;
    image_third: string;
    image_four: string;
    offer_title: string;
    offer_explaination: string;
    offer_cost: number;
    offer_display_date: number;
    offer_expired_date: number;
    offer_deliver_date: number;
    company_id: number;
    offer_count: number;
}